Refactor graph ast/pretty into modules
