+++ partiql-ast/src/visit.rs
@@ -514,22 +514,72 @@ pub trait Visitor<'ast> {
         Traverse::Continue
     }
 
-    fn enter_graph_match(&mut self, _graph_pattern: &'ast ast::GraphMatch) -> Traverse {
+    fn enter_graph_table(&mut self, _gtable: &'ast ast::GraphTable) -> Traverse {
         Traverse::Continue
     }
-    fn exit_graph_match(&mut self, _graph_pattern: &'ast ast::GraphMatch) -> Traverse {
+    fn exit_graph_table(&mut self, _gtable: &'ast ast::GraphTable) -> Traverse {
         Traverse::Continue
     }
 
-    fn enter_graph_match_pattern(
+    fn enter_graph_match(&mut self, _gmatch: &'ast ast::GraphMatch) -> Traverse {
+        Traverse::Continue
+    }
+    fn exit_graph_match(&mut self, _gmatch: &'ast ast::GraphMatch) -> Traverse {
+        Traverse::Continue
+    }
+
+    fn enter_graph_pattern(&mut self, _graph_pattern: &'ast ast::GraphPattern) -> Traverse {
+        Traverse::Continue
+    }
+    fn exit_graph_pattern(&mut self, _graph_pattern: &'ast ast::GraphPattern) -> Traverse {
+        Traverse::Continue
+    }
+
+    fn enter_graph_path_pattern(
         &mut self,
         _graph_pattern: &'ast ast::GraphPathPattern,
     ) -> Traverse {
         Traverse::Continue
     }
-    fn exit_graph_match_pattern(
+    fn exit_graph_path_pattern(&mut self, _graph_pattern: &'ast ast::GraphPathPattern) -> Traverse {
+        Traverse::Continue
+    }
+
+    fn enter_graph_path_sub_pattern(
         &mut self,
-        _graph_pattern: &'ast ast::GraphPathPattern,
+        _graph_pattern: &'ast ast::GraphPathSubPattern,
+    ) -> Traverse {
+        Traverse::Continue
+    }
+    fn exit_graph_path_sub_pattern(
+        &mut self,
+        _graph_pattern: &'ast ast::GraphPathSubPattern,
+    ) -> Traverse {
+        Traverse::Continue
+    }
+
+    fn enter_graph_match_path_pattern(
+        &mut self,
+        _graph_pattern: &'ast ast::GraphMatchPathPattern,
+    ) -> Traverse {
+        Traverse::Continue
+    }
+    fn exit_graph_match_path_pattern(
+        &mut self,
+        _graph_pattern: &'ast ast::GraphMatchPathPattern,
+    ) -> Traverse {
+        Traverse::Continue
+    }
+
+    fn enter_graph_match_path_pattern_quantified(
+        &mut self,
+        _graph_pattern: &'ast ast::GraphMatchPathPatternQuantified,
+    ) -> Traverse {
+        Traverse::Continue
+    }
+    fn exit_graph_match_path_pattern_quantified(
+        &mut self,
+        _graph_pattern: &'ast ast::GraphMatchPathPatternQuantified,
     ) -> Traverse {
         Traverse::Continue
     }
@@ -543,20 +593,6 @@ pub trait Visitor<'ast> {
         Traverse::Continue
     }
     fn exit_graph_match_edge(&mut self, _graph_pattern: &'ast ast::GraphMatchEdge) -> Traverse {
-        Traverse::Continue
-    }
-
-    //TODO
-    fn enter_graph_match_element(
-        &mut self,
-        _graph_pattern: &'ast ast::GraphMatchElement,
-    ) -> Traverse {
-        Traverse::Continue
-    }
-    fn exit_graph_match_element(
-        &mut self,
-        _graph_pattern: &'ast ast::GraphMatchElement,
-    ) -> Traverse {
         Traverse::Continue
     }
